Fix widget blinking
Revert activewidget back to unsigned variable and do not use it for detecting what widget is under edit at the moment. Instead made "hasFocus" public in the renderer. This way it is possible to clear correct widget for blinking. Otherwise there were problem when there are multiple rows with widgets on the screen, first widget of all rows would blink.
So got this sorted now, how to differentiate the widget that is currently in the focus for blinking.